+++ CHANGELOG.md
@@ -5,6 +5,9 @@ of Concierge. Please check the Wiki entry on the release process to understand
 how this file is formatted and how the process works.
 
 ## Unreleased
+### Added
+- Rentals United: add number_of_bathrooms, number single, double, sofa beds fields
+
 ### Fixed
 - Ensure synchronisation.finish! for Kigo host with empty properties
 - Kigo legacy hosts should not be deleted from Concierge